# docs/swagger.yaml
basePath: /
definitions:
...
  app.ListRes:
    properties:
      list:
        description: Data list // 数据清单
      pager:
        allOf:
        - $ref: '#/definitions/app.Pager'
        description: Pagination info // 翻页信息
    type: object
  app.Pager:
    properties:
      page:
        description: Page number // 页码
        type: integer
      pageSize:
        description: Page size // 每页数量
        type: integer
      totalRows:
        description: Total rows // 总行数
        type: integer
    type: object
  app.Res:
    properties:
      code:
        type: integer
      context: {}
      data: {}
      details: {}
      message: {}
      pageIndex:
        description: '1-BASED wire value: 0/absent = non-paginated message, n>0 =
          download page n-1 (internal 0-based). 1-based so page 0 stays distinguishable
          from non-paginated messages under proto3 zero-value elision; only set for
          paginated download messages on pv>=2 connections. Client-to-server PageAck.pageIndex
          stays 0-based. // 线上值为 1-BASED：0/缺省 = 非分页消息，n>0 = 下载页 n-1（内部 0-based）。采用
          1-based 是为了让第 0 页在 proto3 零值不编码规则下仍可与非分页消息区分；仅 pv>=2 连接的分页下载消息会设置。客户端→服务端的
          PageAck.pageIndex 保持 0-based'
      path: {}
      status:
        type: boolean
      vault: {}
    type: object
...
  domain.AutomationEventType:
    enum:
    - cron
    - note_content
    - file_behavior
    - todo_reminder
    - manual
    type: string
    x-enum-varnames:
    - AutomationEventCron
    - AutomationEventNoteContent
    - AutomationEventFileBehavior
    - AutomationEventTodoReminder
    - AutomationEventManual
  domain.AutomationExecutionStatus:
    enum:
    - pending
    - running
    - succeeded
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - AutomationExecutionPending
    - AutomationExecutionRunning
    - AutomationExecutionSucceeded
    - AutomationExecutionFailed
    - AutomationExecutionCancelled
...
  dto.AutomationActionExecutionDTO:
    properties:
      configId:
        type: integer
      error:
        type: string
      finishedAt:
        type: string
      startedAt:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  dto.AutomationExecutionDTO:
    properties:
      actions:
        items:
          $ref: '#/definitions/dto.AutomationActionExecutionDTO'
        type: array
      createdAt:
        type: string
      error:
        type: string
      eventId:
        type: string
      eventType:
        $ref: '#/definitions/domain.AutomationEventType'
      finishedAt:
        type: string
      id:
        type: integer
      startedAt:
        type: string
      status:
        $ref: '#/definitions/domain.AutomationExecutionStatus'
      triggerId:
        type: integer
      uid:
        type: integer
      updatedAt:
        type: string
      vaultId:
        type: integer
    type: object
  dto.AutomationExecutionRetryRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
host: localhost:9000
info:
  contact:
    email: haierkeys@gmail.com
    name: Haierkeys
    url: https://github.com/haierkeys
  description: This is the Fast Note Sync Service HTTP API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Fast Note Sync Service HTTP API
  version: "1.0"
paths:
...
  /api/automations/executions:
    get:
      parameters:
      - in: query
        name: triggerId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.AutomationExecutionDTO'
                        type: array
                    type: object
              type: object
      security:
      - UserAuthToken: []
      summary: List automation executions
      tags:
      - Automation
  /api/automations/executions/retry:
    post:
      consumes:
      - application/json
      parameters:
      - description: Execution ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AutomationExecutionRetryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Retry automation execution
      tags:
      - Automation
...
securityDefinitions:
  ShareAuthToken:
    in: header
    name: Share-Token
    type: apiKey
  UserAuthToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"